Update TaskManagement data class to use Long for taskId and simplify composable definitions in navigation graph

+++ app/src/main/java/com/example/tudeeapp/presentation/navigation/Screens.kt
@@ -33,6 +33,6 @@ sealed class Screens {
     data class CategoryDetails(val id: Long) : Screens()
 
     @Serializable
-    data class TaskManagement(val taskId: Int? = null) : Screens()
+    data class TaskManagement(val taskId: Long? = null) : Screens()
 
 }
